requirements.txt
transformers
datasets
sklearn
pytorch-crf